Extracting video tests
